+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Features
 - Added long description for PyPI ([#642](https://github.com/neptune-ai/neptune-client/pull/642))
 
+### Fixes
+- Fixed GitPython importing during package preparation ([#647](https://github.com/neptune-ai/neptune-client/pull/647))
+
 ## neptune-client 0.10.3
 
 ### Features